# openapi/visana-api-v1.json
{
  "components" : {
    "responses" : {
      "BadRequestProblem" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description" : "The request is invalid."
      },
...
      "ForbiddenProblem" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description" : "The authenticated actor is not allowed to perform this operation."
      },
      "InternalServerProblem" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description" : "An unexpected error occurred. Use the correlation ID for support."
      },
      "NotFoundProblem" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description" : "The requested resource was not found."
      },
      "UnauthorizedProblem" : {
        "content" : {
          "application/problem+json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description" : "Authentication is required."
      },
...
    },
    "schemas" : {
...
      "CreateOrderLineRequest" : {
        "properties" : {
          "productId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "quantity" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "CreateOrderRequest" : {
        "properties" : {
          "lines" : {
            "items" : {
              "$ref" : "#/components/schemas/CreateOrderLineRequest"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "OrderLineResponse" : {
        "properties" : {
          "lineTotal" : {
            "type" : "number"
          },
          "productCode" : {
            "type" : "string"
          },
          "productId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "productName" : {
            "type" : "string"
          },
          "quantity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "unitPrice" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "OrderResponse" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "lines" : {
            "items" : {
              "$ref" : "#/components/schemas/OrderLineResponse"
            },
            "type" : "array"
          },
          "orderId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "ownerActorId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "total" : {
            "type" : "number"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PageOrderResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/OrderResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ProblemDetail" : {
        "properties" : {
          "correlationId" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "VISANA Plan 3 API",
    "version" : "v1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/v1/orders" : {
      "get" : {
        "operationId" : "listOwnedOrders",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : "createdAt",
            "type" : "string"
          }
        }, {
          "description" : "Optional caller correlation ID. The server generates one when absent or invalid.",
          "in" : "header",
          "name" : "X-Correlation-ID",
          "required" : false,
          "schema" : {
            "maxLength" : 128,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageOrderResponse"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequestProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "$ref" : "#/components/responses/ForbiddenProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "List orders owned by the current actor",
        "tags" : [ "Plan 3 Orders" ]
      },
      "post" : {
        "description" : "Only productId and quantity are accepted. The server snapshots catalog prices; this endpoint does not confirm payment.",
        "operationId" : "createOwnedOrder",
        "parameters" : [ {
          "description" : "Optional caller correlation ID. The server generates one when absent or invalid.",
          "in" : "header",
          "name" : "X-Correlation-ID",
          "required" : false,
          "schema" : {
            "maxLength" : 128,
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "Created",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequestProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "$ref" : "#/components/responses/ForbiddenProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Create an actor-owned order",
        "tags" : [ "Plan 3 Orders" ]
      }
    },
    "/api/v1/orders/{orderId}" : {
      "get" : {
        "operationId" : "getOwnedOrder",
        "parameters" : [ {
          "in" : "path",
          "name" : "orderId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Optional caller correlation ID. The server generates one when absent or invalid.",
          "in" : "header",
          "name" : "X-Correlation-ID",
          "required" : false,
          "schema" : {
            "maxLength" : 128,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OrderResponse"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequestProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/UnauthorizedProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "403" : {
            "$ref" : "#/components/responses/ForbiddenProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "$ref" : "#/components/responses/NotFoundProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerProblem",
            "headers" : {
              "X-Correlation-ID" : {
                "description" : "Correlation ID generated or preserved by the server.",
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Read an owned order",
        "tags" : [ "Plan 3 Orders" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
}